backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.9
pdfplumber==0.10.2
pytesseract==0.3.10
pdf2image==1.16.3
nltk==3.8.1
scikit-learn==1.3.2
pillow==10.1.0
python-dotenv==1.0.0
cors==1.0.1